Animationen funktionieren

+++ MildeSprite/ue101SpriteAnimation/TimeBase.pde
@@ -1,7 +1,7 @@
 class TimeBase {
   private int phases = 7;
   private int dauer = 100; // dauer phase in ms
-  private float startTime = 0;
+  private int startTime = 0;
   
   public TimeBase(int dur, int phs) {
     phases = phs;
@@ -10,7 +10,7 @@ class TimeBase {
   }
   
   public int getPhase() {
-    float time = millis() - startTime;
+    int time = millis() - startTime;
     int p = floor(time / dauer);
     return (p % phases);
   }
